requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
django-dbbackup==4.0.2
django-debug-toolbar==4.2.0
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
hiredis==2.2.3
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
loguru==0.7.2
psycopg==3.1.12
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
rpds-py==0.12.0
ruff==0.1.5
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
